requirements.txt
telethon
PyGithub
newspaper3k
tweepy
